\name{OceanView-package}
\alias{OceanView-package}
\alias{OceanView}
\docType{package}
\title{
  Functions for visualising oceanic data sets and model output.
}
\description{
  Visualisation of oceanic data.
 
}

\author{
  Karline Soetaert
}
%\references{
%% to be filled
%}
\seealso{
  \link{db2cross}, converts a dataset from database format to cross table.

  \link{flowpath}, plots velocities as trajectory plot.

  \link{remap}, \link{transect}, 
  \link{extract}, \link{mapsigma}, \link{transectsigma}, 
  mapping and extracting from 2-D or 3-D data.

  \link{Mcommon}, \link{Mplot}, \link{Msplit},
    functions for plotting matrices.

  \link{quiver2D}, velocities plotted as arrows. 

  \link{vectorplot}, vector velocity plot.
}


\keyword{ package }