\name{Reshaping to a crosstable}
\alias{db2cross}
\title{
  Converts a dataset from database-format to a cross table
}
\description{
  Reshapes data arranged in 3 columns to a ``crosstable'' matrix.
}

\usage{
db2cross (input, row = 1, col = 2, value = 3) }

\arguments{
  \item{input }{A \code{matrix} in \emph{database} format, (x,y,z) .
  }
  \item{row }{Number or name of the column in \code{input}, to be used as 
    rows in the result.
  }
  \item{col }{Number or name of the column in \code{input}, to be used as 
    columns in the result.
  }
  \item{value }{Number or name of the column in \code{input}, to be used as 
    values in the result.
  }
}
\value{
  a list containing:
  \item{x }{ The values of the \emph{rows}    }
  \item{y }{ The values of the \emph{columns} }
  \item{z }{ The crosstable, a matrix} 
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}

\examples{
## =======================================================================
## test the function on a small data set
## =======================================================================

 df3 <- data.frame(school = rep(c("a","b","c"), each = 4), 
                   class = rep(9:10, 6),
                   time = rep(c(1,1,2,2), 3),  
                   score = rnorm(12))
 head(df3)
 db2cross(df3, val = 4)

## =======================================================================
## A larger dataset; requires OceanView.Data
## =======================================================================
\dontrun{
 data (pp.aug2009.db)
 crosstab <- db2cross(pp.aug2009.db)
 crosstab$z[crosstab$z>1000] <- 1000
 crosstab$z[crosstab$z<0]    <- NA

 image2D(z = crosstab$z, x = crosstab$x, y = crosstab$y,
       main = "primary production august 2009 mgC/m2/d", 
       NAcol = "black")
 }
}
\details{
 Uses a simple fortran function.
 
 rows and columns are generated by the \code{unique values} in each 
 x- and y-column.
} 
\seealso{
  \link{reshape}, the official (slow) R-function

  \link{map} to remap a matrix or array to higher or lower resolution
}
\keyword{utility}

