\name{OceanView-package}
\alias{OceanView-package}
\alias{OceanView}
\docType{package}
\title{
  Functions to view global ocean data.
}
\description{
  Visualisation functions for global ocean data and model output.
}

\author{
  Karline Soetaert
}
%\references{
%% to be filled
%}
\seealso{
  \link{db2cross}, for converting a data set from database to crosstab 

  \link{mapxy}, to remap a matrix (higher or lower resolution)

  \link{Nitrate3D}, loin, lat, depth values of nitrate from the 
  NODC World Ocean Atlas 2005.
  
  \link{quiver}, \link{quiver.array}, for plotting velocity fields

  \link{flowpath}, for plotting velocity trajectories
  
  \link{Image}, \link{Image.array}, for an image function to visualise 2-D or 3-D data
}


\keyword{ package }